Resolves various issues with Realms search
By setting translucent to false for the navigation bar the following
issues are resolved:

  * Get rid of shadow that appears during transition between search and
    add realm.
  * Realms do not appear behind status bar when scrolling through table

Unfortunately, this introduced bugs with search using
UISearchController. When setting the navigation bar to translucent, the
search bar would animate out of view. It appears this is an issue that
others have experienced with the UISearchController (among other
issues). The realms search now moves away from UISearchController in
favor of implementing a UISearchBar and the UISearchBarDelegate. We
avoided using deprecated UISearchDisplayController.

Resolves:

https://trello.com/c/mJoO3nPw
https://trello.com/c/TeiBEd7r